Hoist state globals into a context structure
In preparation for multi-monitor support, move the global state for a
single screen into a context structure. This structure will eventually
be created per-screen, and the update engine will know what to do
because it's simply passed a state.

Configuration parameters and global state continue to live as globals,
since they don't vary screen to screen.