Convert screen state into an array
Convert the handling of SnScreen into an array of SnScreens, one
associated with each created window. Upon initialization of a window a
screen will be allocated from a pre-allocated array of screens
(ultimately one per monitor). The screen context is associated with that
window. On timer updates, all screens with a valid window are updated.