slowly pan the landscape
To add a little variety, very slowly move the landscape left (slowly, as
in 1 pixel every 5 minutes). As buildings fall off the left, new buildings
are added to the right. This has to be done slowly enough that the rain
completely wipes out the old building lights in between movement, so
that artifacts don't remain.

I tested this by making the timer more frequent, rain and buildings more
plentiful, and then decreasing the period so I could see watch buildings
fall off the left and reappear to the right.